Fixed more entity relationships and added the DB logic for Goals and Discounts, and how they are related for both micro and long term goals for individuals and group budgets

+++ ubuntu-back-end/src/main/java/co/za/ubuntu/ubuntubackend/dto/AccountSplitDTO.java
@@ -15,7 +15,7 @@ public class AccountSplitDTO extends Account {
 
     Integer accountId;
     String splitType;
-    Double incomeAmount;
+    BigDecimal incomeAmount;
     BigDecimal allocationPercentage;
 
 }
